refactor: change recent orders directory
