+++ src/app/admin/(pages)/dashboard/components/RecentOrders.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Table from "./Table";
+import Table from "../../../(components)/Table";
 
 const RecentOrders = () => {
   const orders = [
